+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Paging for Group Polls [#1157](https://github.com/rokwire/illinois-app/issues/1157).
 ### Changed
 - Reworked refresh token functionality [#1168](https://github.com/rokwire/illinois-app/issues/1168).
 - Optimized Sports service startup, data caching and refreshing [#1196](https://github.com/rokwire/illinois-app/issues/1196).
+### Fixed
+- Issues with updating status for Group Polls [#1157](https://github.com/rokwire/illinois-app/issues/1157).
 
 ## [3.1.10] - 2022-01-10
 ### Fixed